database/requirements_enhanced.txt
# Enhanced Vanta Ledger System Requirements
# ========================================

# Core Database Dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
alembic>=1.12.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
fitz>=0.0.1

# AI/ML & Text Processing
spacy>=3.6.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0

# OCR & Image Processing
opencv-python>=4.8.0
tesseract-ocr>=0.1.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment & Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Logging & Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Additional Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
aiofiles>=23.2.0

# Optional: Advanced Analytics
scipy>=1.11.0
statsmodels>=0.14.0

# Optional: Visualization
bokeh>=3.2.0
dash>=2.14.0

# Optional: Cloud & Storage
boto3>=1.28.0
google-cloud-storage>=2.10.0

# Optional: Monitoring & Observability
sentry-sdk>=1.32.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0 